Update aravind.java
There are some problem in some test cases if we take the numCoin[] as int these are the test cases
 250 24
41 34 46 9 37 32 42 21 7 13 1 24 3 43 2 23 8 45 19 30 29 18 35 11

250 26
8 47 13 24 25 31 32 35 3 19 40 48 1 4 17 38 22 30 33 15 44 46 36 9 20 49

245 26
16 30 9 17 40 13 42 5 25 49 7 23 1 44 4 11 33 12 27 2 38 24 28 32 14 50

The problem get solved if we declare numCoin[] as long.

+++ Dynamic Programming/The Coin Change Problem/aravind.java
@@ -19,7 +19,7 @@ public class Solution {
     }
     
     public static int max(int[] coins, int N) {
-        int[] numCoins = new int[N+1];
+        long[] numCoins = new long[N+1];
         numCoins[0] = 1;
         for(int i = 0; i < coins.length; i++) {
             for(int j = coins[i]; j < numCoins.length; j++) {
